refactor: reconfigure localstack and packaging
