+++ src/test/java/awsspringmessaginglab/sj/IntegrationTest.java
@@ -15,6 +15,8 @@ public class IntegrationTest {
 
     @Container
     protected static final LocalStackContainer LOCAL_STACK_CONTAINER = new LocalStackContainer(
-        DockerImageName.parse("localstack/localstack:3.0"))
-        .withServices(LocalStackContainer.Service.SQS, LocalStackContainer.Service.SNS);
+        DockerImageName.parse("localstack/localstack:3.2.0"))
+        .withServices(LocalStackContainer.Service.SQS, LocalStackContainer.Service.SNS)
+        .withEnv("AWS_ACCESS_KEY_ID", "test")
+        .withEnv("AWS_SECRET_ACCESS_KEY", "test");
 }
